requirements.txt
fastapi
uvicorn
python-multipart
fastapi-socketio